+++ CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to this project will be documented in this file. This projec
 
 ## [Unreleased][]
 ### Features
+- **State**: `extensions list` is the default `extensions` subcommand [#147][]
 - **State**: `extensions create|edit|delete` recognize config options [#158][]
 - **State**: `extensions config` formatting options [#159][]
 - **State**: `extensions run` respects global color/format options [#160][]
 - **State**: delete an extension quietly [#162][]
 
+[#147]: https://github.com/Brickster/git-commands/issues/147
 [#158]: https://github.com/Brickster/git-commands/issues/158
 [#159]: https://github.com/Brickster/git-commands/issues/159
 [#160]: https://github.com/Brickster/git-commands/issues/160